+++ packages/protocol-ping/CHANGELOG.md
@@ -41,6 +41,15 @@
   * devDependencies
     * @libp2p/logger bumped from ^4.0.2 to ^4.0.3
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^3.0.4 to ^4.0.0
+    * @libp2p/interface-internal bumped from ^1.0.5 to ^1.0.6
+  * devDependencies
+    * @libp2p/peer-id-factory bumped from ^4.0.3 to ^4.0.4
+
 ## [1.0.8](https://github.com/libp2p/js-libp2p/compare/ping-v1.0.7...ping-v1.0.8) (2024-01-06)